refactor(commons): adjust price data model imports
