+++ standalone/src/models/commons/price/price-draft/presets/minimal.ts
@@ -1,4 +1,4 @@
-import { MoneyDraft } from '../../../money';
+import { MoneyDraft } from '@/models/commons';
 import PriceDraft from '../builder';
 
 const minimal = () =>
